pyproject.toml
[tool.poetry]
name = "python-scripts"
version = "1.0.0"
description = "Python scripts for performance tests"
authors = ["Jesus Roncal <jesus@wordbox.ai>"]

[[tool.poetry.source]]
name = "gaepypi"
url = "https://juan:wordboxjuan@wordboxdev.uc.r.appspot.com/pypi"
secondary = true

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
google-cloud-firestore = "^2.5.2"
google-auth = "^2.6.6"
typing-extensions = "^4.2.0"
authorizationhandler = "^1.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
